+++ crates/alien-ai-gateway/src/router/mod.rs
@@ -9,8 +9,6 @@ use std::sync::Arc;
 use alien_core::ai_catalog::{self, Protocol};
 use alien_core::Platform;
 use alien_error::{AlienError, Context, IntoAlienError};
-use aws_smithy_eventstream::frame::{DecodedFrame, MessageFrameDecoder};
-use aws_smithy_types::event_stream::Message;
 use axum::{
     body::{Body, Bytes},
     extract::{Path, State},
@@ -19,13 +17,30 @@ use axum::{
     routing::{get, post},
     Router,
 };
-use base64::{engine::general_purpose::STANDARD, Engine};
-use futures::StreamExt;
-use serde_json::{json, Map, Value};
-use tracing::warn;
+use serde_json::{json, Value};
 
 use crate::creds::AmbientCred;
 use crate::error::{ErrorData, Result};
+
+mod bedrock;
+mod eventstream;
+mod foundry;
+mod vertex;
+
+use bedrock::proxy_bedrock_anthropic;
+use foundry::proxy_foundry_anthropic;
+use vertex::proxy_vertex_anthropic;
+
+// Re-exported so availability.rs and the test module below can resolve these per-provider
+// items. `vertex_host` is also used by `upstream_target` here; `ensure_block_content` and
+// `EventStreamToSse` are test-only.
+pub(crate) use bedrock::bedrock_geo;
+pub(crate) use foundry::FOUNDRY_ANTHROPIC_VERSION;
+pub(crate) use vertex::vertex_host;
+#[cfg(test)]
+pub(crate) use bedrock::ensure_block_content;
+#[cfg(test)]
+pub(crate) use eventstream::EventStreamToSse;
 
 /// One binding resolved into everything the proxy needs to serve it: the cloud (for
 /// catalog filtering and upstream selection), the location fields used to build the
@@ -388,589 +403,6 @@ fn parse_stream_flag(value: Option<Value>) -> Result<bool> {
     }
 }
 
-/// The Vertex AI Platform host for a location: the global endpoint is the
-/// un-prefixed host; a region prefixes it. The path carries `locations/{location}`
-/// either way.
-pub(crate) fn vertex_host(location: &str) -> String {
-    if location == "global" {
-        "https://aiplatform.googleapis.com".to_string()
-    } else {
-        format!("https://{location}-aiplatform.googleapis.com")
-    }
-}
-
-/// Serve a Claude request through Vertex `rawPredict`. Nearly the Anthropic
-/// Messages API: the model id travels in the URL, streaming picks the URL verb
-/// (`:streamRawPredict`), and the body carries Vertex's version marker instead of
-/// a `model`. The reply is native Anthropic JSON/SSE, so unlike the Bedrock shim
-/// there is no event-stream decoder — and betas ride the standard `anthropic-beta`
-/// header rather than a body field, since Vertex speaks the native Messages API.
-async fn proxy_vertex_anthropic(
-    client: &reqwest::Client,
-    route: &GatewayRoute,
-    upstream_id: &str,
-    mut payload: Value,
-    headers: &HeaderMap,
-) -> Result<Response> {
-    let location = route.region.as_deref().ok_or_else(|| missing_field(route, "location"))?;
-    let project = route.project.as_deref().ok_or_else(|| missing_field(route, "project"))?;
-
-    let obj = payload.as_object_mut().ok_or_else(|| {
-        AlienError::new(ErrorData::InvalidRequest {
-            message: "request body must be a JSON object".to_string(),
-        })
-    })?;
-    obj.remove("model");
-    obj.insert("anthropic_version".to_string(), json!("vertex-2023-10-16"));
-    // The `stream` field stays in the body; Vertex accepts it alongside the verb.
-    let stream = parse_stream_flag(obj.get("stream").cloned())?;
-    let verb = if stream { "streamRawPredict" } else { "rawPredict" };
-
-    let base = route.upstream_base_override.clone().unwrap_or_else(|| vertex_host(location));
-    let url = format!(
-        "{}/v1/projects/{project}/locations/{location}/publishers/anthropic/models/{upstream_id}:{verb}",
-        base.trim_end_matches('/')
-    );
-
-    let upstream_body = serde_json::to_vec(&payload)
-        .into_alien_error()
-        .context(ErrorData::Other {
-            message: "could not re-serialize the rewritten request body".to_string(),
-        })?;
-
-    // Vertex is the native Messages API, so betas ride the standard header —
-    // filtered through the same allowlist that keeps Anthropic-API-side markers
-    // (notably oauth-2025-04-20) from turning the request into a 400.
-    let betas = filtered_header_betas(headers).join(",");
-    let mut extra_headers: Vec<(&str, &str)> = Vec::new();
-    if !betas.is_empty() {
-        extra_headers.push(("anthropic-beta", betas.as_str()));
-    }
-    let upstream =
-        sign_and_execute(client, &route.cred, &url, "", upstream_body, &extra_headers).await?;
-    forward_response(upstream)
-}
-
-/// The Messages API version the gateway bridges to Foundry's Anthropic endpoint;
-/// Foundry reads it from the standard `anthropic-version` header.
-pub(crate) const FOUNDRY_ANTHROPIC_VERSION: &str = "2023-06-01";
-
-/// Serve a Claude request through Foundry's Anthropic endpoint. The closest arm to
-/// the plain passthrough: the model stays in the body (rewritten to the Foundry
-/// deployment name), streaming is the standard body field, and the reply is native
-/// Anthropic JSON/SSE — only the version header and the `/anthropic/v1` path
-/// distinguish it from the OpenAI arm.
-async fn proxy_foundry_anthropic(
-    client: &reqwest::Client,
-    route: &GatewayRoute,
-    upstream_id: &str,
-    mut payload: Value,
-    headers: &HeaderMap,
-) -> Result<Response> {
-    let endpoint =
-        route.azure_endpoint.as_deref().ok_or_else(|| missing_field(route, "endpoint"))?;
-
-    payload["model"] = Value::String(upstream_id.to_string());
-    let upstream_body = serde_json::to_vec(&payload)
-        .into_alien_error()
-        .context(ErrorData::Other {
-            message: "could not re-serialize the rewritten request body".to_string(),
-        })?;
-
-    // The binding carries the AIServices account endpoint; the Anthropic path
-    // serves on that account. Whether the account host also needs the Entra
-    // audience swapped to https://ai.azure.com is settled by the live Foundry
-    // probe — the credential keeps the account audience until that probe says
-    // otherwise.
-    let base = route
-        .upstream_base_override
-        .clone()
-        .unwrap_or_else(|| endpoint.to_string());
-    let url = format!("{}/anthropic/v1/messages", base.trim_end_matches('/'));
-
-    // Foundry speaks the standard Anthropic protocol, so betas ride the standard
-    // header — filtered through the same allowlist that keeps Anthropic-API-side
-    // markers from turning the request into a 400.
-    let betas = filtered_header_betas(headers).join(",");
-    let mut extra_headers = vec![("anthropic-version", FOUNDRY_ANTHROPIC_VERSION)];
-    if !betas.is_empty() {
-        extra_headers.push(("anthropic-beta", betas.as_str()));
-    }
-    let upstream =
-        sign_and_execute(client, &route.cred, &url, "", upstream_body, &extra_headers).await?;
-    forward_response(upstream)
-}
-
-/// The client-executed tool families Bedrock hosts on classic `InvokeModel`
-/// (verified against AWS docs). Anything else typed is server-executed by Anthropic's
-/// own API servers, which Bedrock is not, so it is dropped rather than 400'd.
-const BEDROCK_HOSTED_TOOL_PREFIXES: &[&str] =
-    &["bash_", "text_editor_", "computer_", "memory_", "tool_search_"];
-
-/// The `anthropic_beta` families Bedrock's classic `InvokeModel` accepts in the body
-/// (each live-verified: an accepted tag returns 200, an unknown one is a
-/// ValidationException "invalid beta flag"). Anthropic-API-side markers — notably
-/// `oauth-2025-04-20`, which every OAuth-authenticated Claude Code request declares —
-/// are rejected, so the header bridge folds only these families across.
-const BEDROCK_BETA_PREFIXES: &[&str] = &[
-    "claude-code-",
-    "computer-use-",
-    "context-1m-",
-    "context-management-",
-    "fine-grained-tool-streaming-",
-    "interleaved-thinking-",
-    "output-128k-",
-    "token-efficient-tools-",
-    "tool-examples-",
-];
-
-/// Serve a Claude request through classic Bedrock `InvokeModel`. The Anthropic
-/// Messages body *is* the InvokeModel body, but the model id travels in the URL
-/// (as a cross-region inference profile) and streaming is chosen by the URL
-/// suffix — so, unlike the passthrough path, the body carries neither, and the
-/// streamed reply arrives as AWS event-stream framing we decode back into the
-/// Anthropic SSE the client expects.
-///
-/// This whole function is a protocol shim, kept only until Bedrock's mantle
-/// endpoint serves Claude on the same standard model access InvokeModel already
-/// grants. Where mantle does not yet serve Claude for a given account/region it
-/// returns 403 while the same model and credential serve 200 via InvokeModel;
-/// this shim bridges that gap. Drop it (and the decoder below) in favor of the
-/// plain mantle passthrough once mantle serves Claude directly — to check whether
-/// a region already does:
-///
-/// ```text
-/// curl --aws-sigv4 "aws:amz:<region>:bedrock-mantle" --user "$KEY:$SECRET" \
-///   -H "x-amz-security-token: $TOKEN" -H "content-type: application/json" \
-///   -d '{"model":"anthropic.claude-haiku-4-5","max_tokens":16,
-///        "messages":[{"role":"user","content":"Say ok"}]}' \
-///   https://bedrock-mantle.<region>.api.aws/anthropic/v1/messages
-/// ```
-///
-/// Bedrock's Converse API was evaluated (live, 2026-07-16) and rejected: it
-/// works on standard access, but it speaks AWS's own schema in both directions,
-/// so it would replace these targeted fixups with a full Anthropic⇄Converse
-/// codec (content taxonomy, toolSpec, synthesized message ids, a Converse-event
-/// stream translation) while still needing the event-stream decoder, the system
-/// fold, the server-tool filter, and the beta bridge in relocated form.
-async fn proxy_bedrock_anthropic(
-    client: &reqwest::Client,
-    route: &GatewayRoute,
-    upstream_id: &str,
-    mut payload: Value,
-    headers: &HeaderMap,
-) -> Result<Response> {
-    let region = route.region.as_deref().ok_or_else(|| missing_field(route, "region"))?;
-
-    let obj = payload.as_object_mut().ok_or_else(|| {
-        AlienError::new(ErrorData::InvalidRequest {
-            message: "request body must be a JSON object".to_string(),
-        })
-    })?;
-    // The model is in the URL and streaming is chosen by the URL suffix, so neither
-    // belongs in the body; Bedrock requires its own version marker there instead.
-    obj.remove("model");
-    // Bedrock's schema rejects a body `stream` field, so it is removed here.
-    let stream = parse_stream_flag(obj.remove("stream"))?;
-    obj.insert("anthropic_version".to_string(), json!("bedrock-2023-05-31"));
-
-    // Claude clients declare betas in the `anthropic-beta` HTTP header, but classic
-    // InvokeModel reads only the body's `anthropic_beta`. Bridge the Bedrock-known
-    // families across so a beta-gated tool we forward (computer_*, memory_*) arrives
-    // with the beta it needs — and drop the rest, which Bedrock's body validation
-    // rejects (see merge_beta_headers).
-    merge_beta_headers(obj, headers);
-
-    // Bedrock's InvokeModel schema (pinned to `bedrock-2023-05-31`) predates the
-    // newest Anthropic Messages fields, so a latest client (Claude Code) sends fields
-    // it rejects. Drop the ones outside its schema so the request isn't a 400 — the
-    // gateway is bridging a protocol-version gap, not the raw native endpoint.
-    obj.remove("output_config");
-    obj.remove("context_management");
-    // Bedrock supports only `enabled`/`disabled` extended thinking; drop a newer mode
-    // (e.g. `adaptive`) rather than let Bedrock reject the whole request.
-    let thinking_unsupported = obj
-        .get("thinking")
-        .and_then(|t| t.get("type"))
-        .and_then(Value::as_str)
-        .is_some_and(|t| t != "enabled" && t != "disabled");
-    if thinking_unsupported {
-        obj.remove("thinking");
-    }
-    // Anthropic *server*-executed tool types (web_search, code_execution, web fetch,
-    // advisor) run on Anthropic's own API servers, which InvokeModel is not, so
-    // Bedrock rejects them; drop those and keep the families Bedrock hosts. Dropping
-    // (rather than a 400) is deliberate bridge behavior — Claude Code declares web
-    // tools by default, so rejecting would fail its every request — but it must stay
-    // visible in the logs and leave the body coherent: Bedrock rejects a `tool_choice`
-    // that forces a tool that is no longer declared, and an emptied `tools` array.
-    // A kept beta-gated family still needs its `anthropic_beta` entry, which
-    // merge_beta_headers has already bridged from the header above. Also strip
-    // `defer_loading`, a client-tool field Claude Code's on-demand tool loading adds
-    // that the pinned schema rejects as an extra input.
-    //
-    // Residue blocks a *previous* server-tool turn left in `messages` (e.g.
-    // `web_search_tool_result` from a conversation started on Anthropic's API) are
-    // deliberately NOT rewritten: Bedrock knows those block types and rejects foreign
-    // ones loudly (live-verified: Anthropic-issued `encrypted_content` fails its
-    // validation), and that 400 reaches the client via forward_response — a loud,
-    // honest failure, where stripping would silently alter the conversation.
-    let mut dropped_tools: Vec<String> = Vec::new();
-    let mut tools_remaining = true;
-    if let Some(tools) = obj.get_mut("tools").and_then(Value::as_array_mut) {
-        tools.retain(|tool| {
-            let keep = match tool.get("type").and_then(Value::as_str) {
-                // Plain client tools carry no type, or `custom`.
-                None | Some("custom") => true,
-                Some(tag) => BEDROCK_HOSTED_TOOL_PREFIXES.iter().any(|p| tag.starts_with(p)),
-            };
-            if !keep {
-                let label = tool
-                    .get("name")
-                    .or_else(|| tool.get("type"))
-                    .and_then(Value::as_str)
-                    .unwrap_or("unnamed");
-                dropped_tools.push(label.to_string());
-            }
-            keep
-        });
-        for tool in tools.iter_mut() {
-            if let Some(obj) = tool.as_object_mut() {
-                obj.remove("defer_loading");
-            }
-        }
-        tools_remaining = !tools.is_empty();
-    }
-    if !dropped_tools.is_empty() {
-        warn!(
-            binding = %route.name,
-            tools = %dropped_tools.join(", "),
-            "dropped Anthropic server-executed tools that Bedrock InvokeModel cannot serve"
-        );
-        if !tools_remaining {
-            obj.remove("tools");
-            obj.remove("tool_choice");
-        } else {
-            let forces_dropped = obj
-                .get("tool_choice")
-                .and_then(|choice| choice.get("name"))
-                .and_then(Value::as_str)
-                .is_some_and(|name| dropped_tools.iter().any(|dropped| dropped == name));
-            if forces_dropped {
-                obj.remove("tool_choice");
-            }
-        }
-    }
-    // The pinned schema also predates mid-conversation `system` roles inside
-    // `messages` (top-level `system` is its only sanctioned spot) and enforces
-    // user/assistant alternation. Re-tag those turns as `user` where they stand —
-    // their position in the conversation is what carries the meaning — then fold
-    // same-role neighbors into one message so alternation still holds.
-    if let Some(messages) = obj.get_mut("messages").and_then(Value::as_array_mut) {
-        let originals = std::mem::take(messages);
-        for mut message in originals {
-            if message.get("role").and_then(Value::as_str) == Some("system") {
-                message["role"] = json!("user");
-            }
-            let same_role_as_last = messages
-                .last()
-                .and_then(|previous| previous.get("role"))
-                .is_some_and(|role| Some(role) == message.get("role"));
-            if let Some(previous) = messages.last_mut().filter(|_| same_role_as_last) {
-                let addition = take_content_blocks(&mut message)?;
-                let merged = ensure_block_content(previous)?;
-                // A tool_use turn must be answered by tool_result blocks at the START
-                // of the next message (live-verified: Bedrock 400s on `[text,
-                // tool_result]`), so when the folded-in neighbor carries results —
-                // e.g. a downgraded system turn landed between a tool call and its
-                // result — they slot in right after any results already leading.
-                let (tool_results, rest): (Vec<Value>, Vec<Value>) =
-                    addition.into_iter().partition(|block| {
-                        block.get("type").and_then(Value::as_str) == Some("tool_result")
-                    });
-                let leading = merged
-                    .iter()
-                    .take_while(|block| {
-                        block.get("type").and_then(Value::as_str) == Some("tool_result")
-                    })
-                    .count();
-                merged.splice(leading..leading, tool_results);
-                merged.extend(rest);
-            } else {
-                messages.push(message);
-            }
-        }
-    }
-
-    let upstream_body = serde_json::to_vec(&payload)
-        .into_alien_error()
-        .context(ErrorData::Other {
-            message: "could not re-serialize the Bedrock request body".to_string(),
-        })?;
-
-    let suffix = if stream { "invoke-with-response-stream" } else { "invoke" };
-    let model_id = format!("{}.{}", bedrock_geo(region), upstream_id);
-    let base = route
-        .upstream_base_override
-        .clone()
-        .unwrap_or_else(|| format!("https://bedrock-runtime.{region}.amazonaws.com"));
-    let url = format!("{}/model/{}/{}", base.trim_end_matches('/'), model_id, suffix);
-
-    let upstream = sign_and_execute(client, &route.cred, &url, "bedrock", upstream_body, &[]).await?;
-
-    let status =
-        StatusCode::from_u16(upstream.status().as_u16()).unwrap_or(StatusCode::BAD_GATEWAY);
-
-    // Only a 2xx streaming reply is event-stream framed. A non-2xx (throttling,
-    // missing model access, bad request) and every non-streaming reply are plain
-    // JSON, so forward them untouched — the client sees the real status and body.
-    if !upstream.status().is_success() || !stream {
-        return forward_response(upstream);
-    }
-
-    // Decode the event-stream frames into Anthropic SSE as they arrive (the decoder
-    // buffers across network chunks, so partial frames don't corrupt the output),
-    // then flush once the upstream closes: a stream that ended mid-frame surfaces a
-    // loud error via finish() instead of a silently truncated reply.
-    let sse = futures::stream::unfold(
-        (Box::pin(upstream.bytes_stream()), EventStreamToSse::default(), false),
-        |(mut body, mut decoder, done)| async move {
-            if done {
-                return None;
-            }
-            match body.next().await {
-                Some(Ok(bytes)) => {
-                    Some((Ok(Bytes::from(decoder.push(&bytes))), (body, decoder, false)))
-                }
-                Some(Err(err)) => Some((Err(err), (body, decoder, true))),
-                // Upstream closed: emit the end-of-stream flush, then stop.
-                None => Some((Ok(Bytes::from(decoder.finish())), (body, decoder, true))),
-            }
-        },
-    );
-    Response::builder()
-        .status(status)
-        .header(header::CONTENT_TYPE, "text/event-stream")
-        .body(Body::from_stream(sse))
-        .into_alien_error()
-        .context(ErrorData::Other {
-            message: "could not build the streamed response".to_string(),
-        })
-}
-
-/// Merge the request's `anthropic-beta` headers into the body's `anthropic_beta`.
-/// The header may repeat and each value may be comma-separated; the body field takes
-/// an array or a single string. Body entries are kept and duplicates dropped, so a
-/// client may declare betas any of those ways.
-///
-/// Only BEDROCK_BETA_PREFIXES families are bridged. Bedrock ignores unknown tags in
-/// the *header* but validates the *body* list, so folding an Anthropic-API-side
-/// marker across turns the whole request into a ValidationException. Body entries
-/// stay unfiltered: a client authoring Bedrock-dialect JSON asked for exactly that
-/// list and gets Bedrock's loud answer.
-fn merge_beta_headers(obj: &mut Map<String, Value>, headers: &HeaderMap) {
-    let mut betas: Vec<String> = match obj.get("anthropic_beta") {
-        Some(Value::Array(list)) => {
-            list.iter().filter_map(|v| v.as_str().map(str::to_owned)).collect()
-        }
-        Some(Value::String(tag)) => vec![tag.clone()],
-        _ => Vec::new(),
-    };
-    for beta in filtered_header_betas(headers) {
-        if !betas.iter().any(|existing| existing == &beta) {
-            betas.push(beta);
-        }
-    }
-    if !betas.is_empty() {
-        obj.insert("anthropic_beta".to_string(), json!(betas));
-    }
-}
-
-/// The client's `anthropic-beta` declarations that pass the allowlist. The header
-/// may repeat and each value may be comma-separated. Filtering is an allowlist
-/// because these endpoints validate what they are handed: an Anthropic-API-side
-/// marker (notably `oauth-2025-04-20`, declared by every OAuth Claude Code
-/// request) turns the whole request into a 400. Vertex and Foundry reuse the
-/// Bedrock-verified families until their own live probes verify per-upstream
-/// lists.
-fn filtered_header_betas(headers: &HeaderMap) -> Vec<String> {
-    let mut kept: Vec<String> = Vec::new();
-    let mut dropped: Vec<String> = Vec::new();
-    for value in headers.get_all("anthropic-beta") {
-        let Ok(raw) = value.to_str() else { continue };
-        for beta in raw.split(',').map(str::trim).filter(|b| !b.is_empty()) {
-            if !BEDROCK_BETA_PREFIXES.iter().any(|p| beta.starts_with(p)) {
-                dropped.push(beta.to_string());
-                continue;
-            }
-            if !kept.iter().any(|existing| existing == beta) {
-                kept.push(beta.to_string());
-            }
-        }
-    }
-    if !dropped.is_empty() {
-        warn!(
-            betas = %dropped.join(", "),
-            "dropped anthropic-beta tags outside the allowlisted families"
-        );
-    }
-    kept
-}
-
-/// Normalize a message's `content` to a block array and hand the array back, so two
-/// messages folding into one can concatenate their block lists. A string becomes a
-/// single text block; an existing array is kept; any other shape (missing, null, an
-/// object) is a malformed message the native Anthropic endpoint would reject — fail
-/// loud rather than fold the turn into an empty array and answer a conversation the
-/// client didn't send.
-fn ensure_block_content(message: &mut Value) -> Result<&mut Vec<Value>> {
-    match message.get("content") {
-        Some(Value::Array(_)) => {}
-        Some(Value::String(text)) => {
-            message["content"] = json!([{ "type": "text", "text": text }]);
-        }
-        _ => {
-            return Err(AlienError::new(ErrorData::InvalidRequest {
-                message: "every message `content` must be a string or an array of blocks"
-                    .to_string(),
-            }))
-        }
-    }
-    Ok(message["content"]
-        .as_array_mut()
-        .expect("content was just normalized to an array"))
-}
-
-/// Take a message's content as a block list, leaving the message with an empty one.
-fn take_content_blocks(message: &mut Value) -> Result<Vec<Value>> {
-    Ok(std::mem::take(ensure_block_content(message)?))
-}
-
-/// The cross-region inference-profile geo prefix for a Bedrock region. Claude on
-/// Bedrock is invocable only through a geo profile (e.g. `us.anthropic.…`).
-///
-/// us / us-gov regions keep their own geo. Every other commercial region routes via
-/// the region-agnostic `global` profile: current-generation Claude models publish a
-/// `global.` inference profile invocable from any commercial region (verified against
-/// live Bedrock), and do NOT publish `eu.`/`apac.` profiles, so a per-continent
-/// prefix would build a non-existent id. (An older model that publishes only a `us.`
-/// profile, e.g. opus-4.1, stays us-region-only either way.)
-pub(crate) fn bedrock_geo(region: &str) -> &'static str {
-    if region.starts_with("us-gov-") {
-        "us-gov"
-    } else if region.starts_with("us-") {
-        "us"
-    } else {
-        "global"
-    }
-}
-
-/// Decoder turning Bedrock's `vnd.amazon.eventstream` framing into Anthropic SSE.
-/// A normal chunk frame's payload is `{"bytes": base64(<anthropic event json>)}`;
-/// the decoded event carries a `type` we surface as the SSE `event:` name. Network
-/// chunks can split or merge frames, so bytes are buffered until each frame is
-/// whole. Frame parsing (prelude, headers, both CRC32 checks) is
-/// aws-smithy-eventstream's `MessageFrameDecoder` — AWS's own decoder for this
-/// wire format — so a corrupted or desynced stream fails its CRCs instead of
-/// decoding to garbage.
-#[derive(Default)]
-struct EventStreamToSse {
-    buf: Vec<u8>,
-    /// Set once a frame fails to decode (a CRC mismatch or malformed prelude).
-    /// From then on the buffer can never be drained, so we stop parsing rather than
-    /// spin on it forever.
-    failed: bool,
-}
-
-impl EventStreamToSse {
-    /// Append a network chunk and return the SSE for every frame it now completes.
-    fn push(&mut self, chunk: &[u8]) -> String {
-        if self.failed {
-            return String::new();
-        }
-        self.buf.extend_from_slice(chunk);
-        let mut out = String::new();
-        // A fresh decoder scans the buffer from the top each push, and only the bytes
-        // of fully decoded frames are drained: MessageFrameDecoder consumes a prelude
-        // into internal state before its frame completes, so reusing one across
-        // pushes would strand those bytes between the two buffers.
-        let mut decoder = MessageFrameDecoder::new();
-        let mut cursor: &[u8] = &self.buf;
-        let mut consumed = 0;
-        loop {
-            match decoder.decode_frame(&mut cursor) {
-                Ok(DecodedFrame::Complete(message)) => {
-                    consumed = self.buf.len() - cursor.len();
-                    out.push_str(&message_to_sse(&message));
-                }
-                Ok(DecodedFrame::Incomplete) => break,
-                Err(_) => {
-                    // A CRC mismatch or malformed prelude can never recover: the byte
-                    // stream desynced. Surface it loudly and stop, rather than
-                    // silently stall on bytes we can never drain (which would
-                    // truncate the reply under an already-sent 200).
-                    self.failed = true;
-                    out.push_str(&error_sse("the model response stream could not be decoded"));
-                    break;
-                }
-            }
-        }
-        self.buf.drain(0..consumed);
-        out
-    }
-
-    /// Flush at end of stream. A non-empty buffer here means the upstream closed
-    /// mid-frame (a truncated or desynced stream), so surface a loud error rather
-    /// than drop the tail; a clean boundary (empty buffer) emits nothing.
-    fn finish(&mut self) -> String {
-        if self.failed || self.buf.is_empty() {
-            return String::new();
-        }
-        self.failed = true;
-        error_sse("the model response stream ended before the final frame completed")
-    }
-}
-
-/// One event-stream message rendered as the Anthropic SSE the client expects.
-///
-/// A normal chunk wraps the event as `{"bytes": base64(...)}`. Anything else on an
-/// InvokeModelWithResponseStream reply is an exception frame: Bedrock signals
-/// mid-stream failures (throttlingException, modelStreamErrorException,
-/// internalServerException) this way, with the exception body as the raw payload.
-/// Such a frame is surfaced as an Anthropic `error` SSE event rather than dropped,
-/// because dropping it would truncate the reply under an already-sent 200 with no
-/// error reaching the client.
-fn message_to_sse(message: &Message) -> String {
-    let outer: Option<Value> = serde_json::from_slice(message.payload()).ok();
-    if let Some(sse) = outer.as_ref().and_then(chunk_to_sse) {
-        return sse;
-    }
-    // Exception / error frame: forward Bedrock's own message so the client sees why.
-    let message = outer
-        .as_ref()
-        .and_then(|o| o.get("message"))
-        .and_then(Value::as_str)
-        .unwrap_or("the model returned an error mid-stream");
-    error_sse(message)
-}
-
-/// A normal `{"bytes": base64(<anthropic event>)}` chunk rendered as its SSE line,
-/// or `None` if the frame is not a well-formed chunk.
-fn chunk_to_sse(outer: &Value) -> Option<String> {
-    let event_bytes = STANDARD.decode(outer.get("bytes")?.as_str()?).ok()?;
-    let event: Value = serde_json::from_slice(&event_bytes).ok()?;
-    let event_type = event.get("type")?.as_str()?;
-    let data = std::str::from_utf8(&event_bytes).ok()?;
-    Some(format!("event: {event_type}\ndata: {data}\n\n"))
-}
-
-/// An Anthropic `error` SSE event carrying `message`, so a mid-stream failure
-/// reaches the client as a loud error instead of a silently truncated reply.
-fn error_sse(message: &str) -> String {
-    let event = json!({ "type": "error", "error": { "type": "api_error", "message": message } });
-    format!("event: error\ndata: {event}\n\n")
-}
-
 #[cfg(test)]
 mod tests {
     use std::net::Ipv4Addr;
@@ -978,6 +410,8 @@ mod tests {
     use aws_credential_types::provider::SharedCredentialsProvider;
     use aws_credential_types::Credentials;
     use aws_smithy_eventstream::frame::write_message_to;
+    use aws_smithy_types::event_stream::Message;
+    use base64::{engine::general_purpose::STANDARD, Engine};
     use httpmock::prelude::*;
 
     use super::*;
